Applied DDD Pattern and Code Refactoring for a better OOP Approach
