+++ src/main/java/com/bsc36/project11cs/application/visualizer/SmartGroup.java
@@ -2,33 +2,30 @@ package com.bsc36.project11cs.application.visualizer;
 
 import javafx.scene.Group;
 import javafx.scene.transform.Rotate;
-import javafx.scene.transform.Transform;
 
 /**
- * SmartGroup class
+ * The SmartGroup class extends JavaFX Group to provide additional rotation capabilities.
+ * It includes methods for rotating the group around the X and Y axes.
  */
 public class SmartGroup extends Group {
-    Rotate r = new Rotate();
-
-    Transform t = new Rotate();
 
     /**
-     * Rotate by X
+     * Rotates the group around the X-axis.
      *
-     * @param ang int
+     * @param angle The angle in degrees by which to rotate the group.
      */
-    public void rotateByX(int ang) {
-        t = new Rotate(ang, Rotate.X_AXIS);
-        this.getTransforms().add(t);
+    public void rotateByX(int angle) {
+        Rotate rotationX = new Rotate(angle, Rotate.X_AXIS);
+        this.getTransforms().add(rotationX);
     }
 
     /**
-     * Rotate by Y
+     * Rotates the group around the Y-axis.
      *
-     * @param ang int
+     * @param angle The angle in degrees by which to rotate the group.
      */
-    public void rotateByY(int ang) {
-        t = new Rotate(ang, Rotate.Y_AXIS);
-        this.getTransforms().add(t);
+    public void rotateByY(int angle) {
+        Rotate rotationY = new Rotate(angle, Rotate.Y_AXIS);
+        this.getTransforms().add(rotationY);
     }
 }
