+++ src/main/java/com/bsc36/project11cs/domain/entities/CargoSpace.java
@@ -1,48 +1,42 @@
 package com.bsc36.project11cs.domain.entities;
 
-import com.bsc36.project11cs.domain.entities.parcel.Parcel;
-import com.bsc36.project11cs.infrastructure.configuration.BasicConfig;
-import com.bsc36.project11cs.application.visualizer.SmartGroup;
-import javafx.scene.paint.Color;
-import javafx.scene.paint.PhongMaterial;
 import javafx.scene.shape.Box;
+import javafx.scene.paint.Color;
 import javafx.scene.shape.DrawMode;
+import javafx.scene.paint.PhongMaterial;
+import com.bsc36.project11cs.domain.valueobjects.Size;
+import com.bsc36.project11cs.domain.entities.parcel.Parcel;
+import com.bsc36.project11cs.application.visualizer.SmartGroup;
+import com.bsc36.project11cs.infrastructure.configuration.BasicConfig;
 
 public class CargoSpace {
     private final SmartGroup space;
     private int[][][] shape;
-    private final int width;
-    private final int height;
-    private final int length;
+    private final Size size;
     private int score;
 
     /**
      * Constructor
-     * @param length Length
-     * @param height Height
-     * @param width Width
+     * @param size Size of the cargo space
      */
-    public CargoSpace(int length, int height, int width) {
-        this.length = length;
-        this.height = height;
-        this.width = width;
-
-        this.shape = new int[length][height][width];
+    public CargoSpace(Size size) {
+        this.size = size;
+        this.shape = new int[size.length()][size.height()][size.width()];
 
         // Initialize the shape
-        for (int x = 0; x < length; x++) {
-            for (int y = 0; y < height; y++) {
-                for (int z = 0; z < width; z++) {
+        for (int x = 0; x < size.length(); x++) {
+            for (int y = 0; y < size.height(); y++) {
+                for (int z = 0; z < size.width(); z++) {
                     this.shape[x][y][z] = 0;
                 }
             }
         }
 
+        // Create the cargo space Box
         Box spaceBox = createSpaceBox();
-
         // Add the cargo space to the SmartGroup
         this.space = new SmartGroup();
-        setupSpace(length, height, width, spaceBox);
+        setupSpace(spaceBox);
     }
 
     /**
@@ -52,7 +46,7 @@ public class CargoSpace {
         this.space.getChildren().clear();
         // Create the cargo space Box
         Box spaceBox = createSpaceBox();
-        setupSpace(this.length, this.height, this.width, spaceBox);
+        setupSpace(spaceBox);
     }
 
     /**
@@ -79,12 +73,22 @@ public class CargoSpace {
         return this.space;
     }
 
+
+
+    /**
+     * Get the length of the cargo space
+     * @return int Length of the cargo space
+     */
+    public int getLength() {
+        return this.size.length();
+    }
+
     /**
      * Get the width of the cargo space
      * @return int Width of the cargo space
      */
     public int getWidth() {
-        return width;
+        return this.size.width();
     }
 
     /**
@@ -92,15 +96,7 @@ public class CargoSpace {
      * @return int Height of the cargo space
      */
     public int getHeight() {
-        return height;
-    }
-
-    /**
-     * Get the length of the cargo space
-     * @return int Length of the cargo space
-     */
-    public int getLength() {
-        return length;
+        return this.size.height();
     }
 
     /**
@@ -112,9 +108,9 @@ public class CargoSpace {
      */
     private Box getSmallBox(int x, int y, int z) {
         Box smallBox = new Box(BasicConfig.INCREASE_RATIO, BasicConfig.INCREASE_RATIO, BasicConfig.INCREASE_RATIO);
-        smallBox.setTranslateX(x * BasicConfig.INCREASE_RATIO - this.length * BasicConfig.INCREASE_RATIO / 2.0 + BasicConfig.INCREASE_RATIO / 2.0);
-        smallBox.setTranslateY(y * BasicConfig.INCREASE_RATIO - this.height * BasicConfig.INCREASE_RATIO / 2.0 + BasicConfig.INCREASE_RATIO / 2.0);
-        smallBox.setTranslateZ(z * BasicConfig.INCREASE_RATIO - this.width * BasicConfig.INCREASE_RATIO / 2.0 + BasicConfig.INCREASE_RATIO / 2.0);
+        smallBox.setTranslateX(x * BasicConfig.INCREASE_RATIO - this.size.length() * BasicConfig.INCREASE_RATIO / 2.0 + BasicConfig.INCREASE_RATIO / 2.0);
+        smallBox.setTranslateY(y * BasicConfig.INCREASE_RATIO - this.size.height() * BasicConfig.INCREASE_RATIO / 2.0 + BasicConfig.INCREASE_RATIO / 2.0);
+        smallBox.setTranslateZ(z * BasicConfig.INCREASE_RATIO - this.size.width() * BasicConfig.INCREASE_RATIO / 2.0 + BasicConfig.INCREASE_RATIO / 2.0);
         // Use different colors for the small boxes based on their position
         if ((x + y + z) % 2 == 0) {
             smallBox.setMaterial(new PhongMaterial(Color.LIGHTGRAY));
@@ -131,7 +127,7 @@ public class CargoSpace {
      * @param y Y coordinate
      * @param z Z coordinate
      */
-    public void addParcel(Parcel parcel, int x, int y, int z) {
+    public void placeParcel(Parcel parcel, int x, int y, int z) {
         int[][][] shape = parcel.getShape();
         // Manage the color by the id
         for (int i = 0; i < shape.length; i++) {
@@ -151,12 +147,12 @@ public class CargoSpace {
     }
 
     /**
-     * Remove a parcel from the cargo space
-     * @param parcel Parcel
-     * @param x X coordinate
-     * @param y Y coordinate
-     * @param z Z coordinate
-     */
+    * Remove a parcel from the cargo space
+    * @param parcel Parcel
+    * @param x X coordinate
+    * @param y Y coordinate
+    * @param z Z coordinate
+    */
     public void removeParcel(Parcel parcel, int x, int y, int z) {
         int[][][] shape = parcel.getShape();
         // Manage the color by the id
@@ -174,31 +170,35 @@ public class CargoSpace {
     }
 
     /**
-     * Get the score of the cargo space
-     * @return int Score of the cargo space
-     */
+    * Get the score of the cargo space
+    * @return int Score of the cargo space
+    */
     public int getScore() {
         return score;
     }
 
     /**
-     * Set the score of the cargo space
-     * @param score Score of the cargo space
-     */
+    * Set the score of the cargo space
+    * @param score Score of the cargo space
+    */
     public void setScore(int score) {
         this.score = score;
     }
 
+    /**
+    * Update the state of the cargo space
+    * @param newArr New array
+    */
     public void updateState(int[][][] newArr) {
         space.getChildren().removeAll(space.getChildren());
         placeBox(newArr);
     }
 
     /**
-     * Place the box
-     * @param newArr New array
-     */
-   private void placeBox(int[][][] newArr) {
+    * Place the box
+    * @param newArr New array
+    */
+    private void placeBox(int[][][] newArr) {
         for (int i = 0; i < newArr.length; i++) {
             for (int j = 0; j < newArr[0].length; j++) {
                 for (int k = 0; k < newArr[0][0].length; k++) {
@@ -212,19 +212,16 @@ public class CargoSpace {
     }
 
     /**
-     * Set up the cargo space
-     * @param length Length of the cargo space
-     * @param height Height of the cargo space
-     * @param width Width of the cargo space
-     * @param spaceBox Box
-     */
-    private void setupSpace(int length, int height, int width, Box spaceBox) {
+    * Set up the cargo space
+    * @param spaceBox Box
+    */
+    private void setupSpace(Box spaceBox) {
         // Add the cargo space Box to the cargo space
         this.space.getChildren().add(spaceBox);
         // Add the small boxes to the cargo space
-        for (int x = 0; x < length; x++) {
-            for (int y = 0; y < height; y++) {
-                for (int z = 0; z < width; z++) {
+        for (int x = 0; x < this.size.length(); x++) {
+            for (int y = 0; y <  this.size.height(); y++) {
+                for (int z = 0; z < this.size.width(); z++) {
                     Box smallBox = getSmallBox(x, y, z);
                     this.space.getChildren().add(smallBox);
                 }
@@ -233,12 +230,12 @@ public class CargoSpace {
     }
 
     /**
-     * Set a box
-     * @param x X coordinate
-     * @param y Y coordinate
-     * @param z Z coordinate
-     * @param id Parcel's ID
-     */
+    * Set a box
+    * @param x X coordinate
+    * @param y Y coordinate
+    * @param z Z coordinate
+    * @param id Parcel's ID
+    */
     private void setBox(int x, int y, int z, int id) {
         for (int i = 0; i < shape.length; i++) {
             for (int j = 0; j < shape[0].length; j++) {
@@ -257,26 +254,25 @@ public class CargoSpace {
     }
 
     /**
-     * Get the parcel from the id
-     * @param id Parcel's ID
-     * @return Parcel
-     */
+    * Get the parcel from the id
+    * @param id Parcel's ID
+    * @return Parcel
+    */
     private Parcel getParcelFromId(int id) {
        return Parcel.createParcelFromId(id);
     }
 
     /**
-     * Create the cargo space Box
-     * @return Box
-     */
+    * Create the cargo space Box
+    * @return Box
+    */
     private Box createSpaceBox() {
         // Create the cargo space Box
         Box spaceBox = new Box(
-                this.length * BasicConfig.INCREASE_RATIO,
-                this.height * BasicConfig.INCREASE_RATIO,
-                this.width * BasicConfig.INCREASE_RATIO
+                this.size.length() * BasicConfig.INCREASE_RATIO,
+                this.size.height() * BasicConfig.INCREASE_RATIO,
+                this.size.width() * BasicConfig.INCREASE_RATIO
         );
-
         // Setting the color of the cargo space
         spaceBox.setMaterial(new PhongMaterial(Color.DARKGRAY));
         spaceBox.setDrawMode(DrawMode.LINE);
