+++ src/main/java/com/bsc36/project11cs/domain/entities/parcel/ParcelC.java
@@ -1,19 +1,28 @@
 package com.bsc36.project11cs.domain.entities.parcel;
 
-import com.bsc36.project11cs.infrastructure.configuration.BasicConfig;
 import javafx.scene.paint.Color;
+import com.bsc36.project11cs.domain.valueobjects.Size;
+import com.bsc36.project11cs.infrastructure.configuration.BasicConfig;
 
+
+/**
+ * The ParcelC class represents a specific type of parcel (Parcel C).
+ * It extends the abstract class Parcel and defines the characteristics
+ * and structure of Parcel C, including its ID, type, name, color, size,
+ * volume, 3D shape, and rotations.
+ */
 public class ParcelC extends Parcel{
 
     /**
-     * Constructor for ParcelC
+     * Constructor for ParcelC initializes the attributes of Parcel C.
      */
     public ParcelC() {
         this.id = 3;
         this.type = 'C';
         this.name = "Parcel C";
         this.color = Color.valueOf(BasicConfig.C_PARCEL_COLOR);
-        this.volume = 27;
+        this.size = new Size(3, 3, 3);
+        this.volume = calculateVolume();
         // 3D shape of the parcel
         this.shape = new int[][][] {
                 {{1, 1, 1}, {1, 1, 1}, {1, 1, 1}},
