+++ src/main/resources/com/bsc36/project11cs/views/parcels-view.fxml
@@ -10,7 +10,7 @@
 <?import javafx.scene.layout.VBox?>
 <?import javafx.scene.text.Font?>
 
-<VBox alignment="CENTER" maxHeight="1.7976931348623157E308" maxWidth="1.7976931348623157E308" spacing="10" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.bsc36.project11cs.application.controllers.ParcelsController">
+<VBox alignment="CENTER" maxHeight="1.7976931348623157E308" maxWidth="1.7976931348623157E308" spacing="10" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.bsc36.project11cs.application.controllers.services.ParcelsController">
    <GridPane alignment="CENTER" hgap="10.0" vgap="10.0">
       <columnConstraints>
          <ColumnConstraints hgrow="ALWAYS" maxWidth="1.7976931348623157E308" />
